+++ src/utils/selectors/QACert/assert.js
@@ -3,6 +3,7 @@ import * as qaApi from "../../api/qaCertificationsAPI";
 // Selectors that normalize api data to fit the columns in UI datatable
 import * as selector from "./TestSummary";
 import * as injectionSelector from "./LinearityInjection";
+import * as cycleTimeInjections from "./CycleTimeInjection";
 // Table Names
 const proGas = "Protocol Gas";
 const lineTest = "Linearity Test";
@@ -236,7 +237,7 @@ export const getDataTableRecords = (dataIn, name) => {
     case cycleTimeSummary:
       return selector.mapCycleTimeSummariesToRows(dataIn);
     case cycleTimeInjection:
-      return selector.mapCycleTimeInjectionsToRows(dataIn);
+      return cycleTimeInjections.mapCycleTimeInjectionsToRows(dataIn);
     case transmitterTransducerAccuracyData:
       return selector.mapTransmitterTransducerAccuracyDataToRows(dataIn);
     case flowToLoadReference:
